Fixed an issue with duplicate heading names
The parent of any sub heading except the first one was not stored,
which resulted in duplicate ids being generated.

The HTML generator has been updated to process the headings in order
to fix an issue where two headings on the same sub level with the
same title would be processed incorrectly.